// frontend/src/shared/api/notes/model/createNoteRequestScope.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * defShows Notes & Recap API
 * OpenAPI spec version: 0.1.0
 */

export type CreateNoteRequestScope = typeof CreateNoteRequestScope[keyof typeof CreateNoteRequestScope];


export const CreateNoteRequestScope = {
  show: 'show',
  season: 'season',
  episode: 'episode',
} as const;